Create generic terraform workflow state notifier.
