+++ server/neptune/workflows/internal/deploy/terraform/state.go
@@ -6,17 +6,16 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/runatlantis/atlantis/server/events/metrics"
-	key "github.com/runatlantis/atlantis/server/neptune/context"
 
 	"github.com/runatlantis/atlantis/server/neptune/workflows/activities"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github/markdown"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/terraform"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/notifier"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/terraform/state"
-	"go.temporal.io/sdk/temporal"
 	"go.temporal.io/sdk/workflow"
 )
+
+type TerraformWorkflowNotifier interface {
+	Notify(workflow.Context, DeploymentInfo, *state.Workflow) error
+}
 
 type auditActivities interface {
 	AuditJob(ctx context.Context, request activities.AuditJobRequest) error
@@ -28,8 +27,8 @@ type receiverActivities interface {
 }
 
 type StateReceiver struct {
-	Activity             receiverActivities
-	CheckRunSessionCache CheckRunClient
+	Activity  receiverActivities
+	Notifiers []TerraformWorkflowNotifier
 }
 
 func (n *StateReceiver) Receive(ctx workflow.Context, c workflow.ReceiveChannel, deploymentInfo DeploymentInfo) {
@@ -40,10 +39,11 @@ func (n *StateReceiver) Receive(ctx workflow.Context, c workflow.ReceiveChannel,
 		metrics.SignalNameTag: state.WorkflowStateChangeSignal,
 	}).Counter(metrics.SignalReceive).Inc(1)
 
-	// TODO: if we never created a check run, there was likely some issue, we should attempt to create it again.
-	if deploymentInfo.CheckRunID == 0 {
-		workflow.GetLogger(ctx).Error("check run id is 0, skipping update of check run")
-		return
+	for _, notifier := range n.Notifiers {
+		if err := notifier.Notify(ctx, deploymentInfo, workflowState); err != nil {
+			workflow.GetLogger(ctx).Error(errors.Wrap(err, "notifying workflow state change").Error())
+		}
+
 	}
 
 	// emit audit events when Apply operation is run
@@ -52,40 +52,6 @@ func (n *StateReceiver) Receive(ctx workflow.Context, c workflow.ReceiveChannel,
 			workflow.GetLogger(ctx).Error(errors.Wrap(err, "auditing apply job event").Error())
 		}
 	}
-
-	if err := n.updateCheckRun(ctx, workflowState, deploymentInfo); err != nil {
-		workflow.GetLogger(ctx).Error("updating check run", key.ErrKey, err)
-	}
-}
-
-func (n *StateReceiver) updateCheckRun(ctx workflow.Context, workflowState *state.Workflow, deploymentInfo DeploymentInfo) error {
-	summary := markdown.RenderWorkflowStateTmpl(workflowState)
-	checkRunState := determineCheckRunState(workflowState)
-
-	request := notifier.GithubCheckRunRequest{
-		Title:   BuildCheckRunTitle(deploymentInfo.Root.Name),
-		Sha:     deploymentInfo.Commit.Revision,
-		State:   checkRunState,
-		Repo:    deploymentInfo.Repo,
-		Summary: summary,
-	}
-
-	if workflowState.Apply != nil {
-		// add any actions pertaining to the apply job
-		for _, a := range workflowState.Apply.GetActions().Actions {
-			request.Actions = append(request.Actions, a.ToGithubCheckRunAction())
-		}
-	}
-
-	// cap our retries for non-terminal states to allow for at least some progress
-	if checkRunState != github.CheckRunFailure && checkRunState != github.CheckRunSuccess {
-		ctx = workflow.WithRetryPolicy(ctx, temporal.RetryPolicy{
-			MaximumAttempts: 3,
-		})
-	}
-
-	_, err := n.CheckRunSessionCache.CreateOrUpdate(ctx, deploymentInfo.ID.String(), request)
-	return err
 }
 
 func (n *StateReceiver) emitApplyEvents(ctx workflow.Context, jobState *state.Job, deploymentInfo DeploymentInfo) error {
@@ -122,37 +88,4 @@ func (n *StateReceiver) emitApplyEvents(ctx workflow.Context, jobState *state.Jo
 	}
 
 	return workflow.ExecuteActivity(ctx, n.Activity.AuditJob, auditJobReq).Get(ctx, nil)
-}
-
-func determineCheckRunState(workflowState *state.Workflow) github.CheckRunState {
-	if waitingForActionOn(workflowState.Plan) || waitingForActionOn(workflowState.Apply) {
-		return github.CheckRunActionRequired
-	}
-
-	if workflowState.Result.Status != state.CompleteWorkflowStatus {
-		return github.CheckRunPending
-	}
-
-	if workflowState.Result.Reason == state.SuccessfulCompletionReason {
-		return github.CheckRunSuccess
-	}
-
-	timeouts := []state.WorkflowCompletionReason{
-		state.TimeoutError,
-		state.ActivityDurationTimeoutError,
-		state.HeartbeatTimeoutError,
-		state.SchedulingTimeoutError,
-	}
-
-	for _, t := range timeouts {
-		if workflowState.Result.Reason == t {
-			return github.CheckRunTimeout
-		}
-	}
-
-	return github.CheckRunFailure
-}
-
-func waitingForActionOn(job *state.Job) bool {
-	return job != nil && job.Status == state.WaitingJobStatus && len(job.OnWaitingActions.Actions) > 0
 }
