+++ server/neptune/workflows/internal/deploy/terraform/state_test.go
@@ -3,20 +3,17 @@ package terraform_test
 import (
 	"context"
 	"net/url"
-	"strconv"
 	"testing"
 	"time"
 
 	"github.com/google/uuid"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/activities"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github"
-	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/github/markdown"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/activities/terraform"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/notifier"
 	internalTerraform "github.com/runatlantis/atlantis/server/neptune/workflows/internal/deploy/terraform"
 	"github.com/runatlantis/atlantis/server/neptune/workflows/internal/terraform/state"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"go.temporal.io/sdk/testsuite"
 	"go.temporal.io/sdk/workflow"
 )
@@ -37,44 +34,63 @@ func (t *testCheckRunClient) CreateOrUpdate(ctx workflow.Context, deploymentID s
 type testActivities struct {
 }
 
-func (a *testActivities) GithubUpdateCheckRun(ctx context.Context, request activities.UpdateCheckRunRequest) (activities.UpdateCheckRunResponse, error) {
-	return activities.UpdateCheckRunResponse{}, nil
-}
-
 func (a *testActivities) AuditJob(ctx context.Context, request activities.AuditJobRequest) error {
 	return nil
 }
 
-type stateReceiveRequest struct {
-	StatesToSend    []*state.Workflow
-	DeploymentInfo  internalTerraform.DeploymentInfo
-	ExpectedRequest notifier.GithubCheckRunRequest
-	T               *testing.T
+type testNotifier struct {
+	expectedInfo  internalTerraform.DeploymentInfo
+	expectedState *state.Workflow
+	expectedT     *testing.T
+	called        bool
 }
 
-func testStateReceiveWorkflow(ctx workflow.Context, r stateReceiveRequest) error {
+func (n *testNotifier) Notify(ctx workflow.Context, info internalTerraform.DeploymentInfo, s *state.Workflow) error {
+	assert.Equal(n.expectedT, n.expectedInfo, info)
+	assert.Equal(n.expectedT, n.expectedState, s)
+
+	n.called = true
+
+	return nil
+}
+
+type stateReceiveRequest struct {
+	State          *state.Workflow
+	DeploymentInfo internalTerraform.DeploymentInfo
+	T              *testing.T
+}
+
+type stateReceiveResponse struct {
+	NotifierCalled bool
+}
+
+func testStateReceiveWorkflow(ctx workflow.Context, r stateReceiveRequest) (stateReceiveResponse, error) {
 	ctx = workflow.WithActivityOptions(ctx, workflow.ActivityOptions{
 		ScheduleToCloseTimeout: 5 * time.Second,
 	})
 	ch := workflow.NewChannel(ctx)
 
+	notifier := &testNotifier{
+		expectedInfo:  r.DeploymentInfo,
+		expectedState: r.State,
+		expectedT:     r.T,
+	}
+
 	receiver := &internalTerraform.StateReceiver{
-		Activity: &testActivities{},
-		CheckRunSessionCache: &testCheckRunClient{
-			expectedRequest: r.ExpectedRequest,
-			expectedT:       r.T,
+		Notifiers: []internalTerraform.TerraformWorkflowNotifier{
+			notifier,
 		},
 	}
 
 	workflow.Go(ctx, func(ctx workflow.Context) {
-		for _, s := range r.StatesToSend {
-			ch.Send(ctx, s)
-		}
+		ch.Send(ctx, r.State)
 	})
 
 	receiver.Receive(ctx, ch, r.DeploymentInfo)
 
-	return nil
+	return stateReceiveResponse{
+		NotifierCalled: notifier.called,
+	}, nil
 }
 
 func TestStateReceive(t *testing.T) {
@@ -85,8 +101,6 @@ func TestStateReceive(t *testing.T) {
 		URL: outputURL,
 	}
 
-	stTime := time.Now()
-	endTime := stTime.Add(time.Second * 5)
 	internalDeploymentInfo := internalTerraform.DeploymentInfo{
 		CheckRunID: 1,
 		ID:         uuid.New(),
@@ -97,263 +111,26 @@ func TestStateReceive(t *testing.T) {
 		},
 	}
 
-	cases := []struct {
-		State                   *state.Workflow
-		ExpectedCheckRunState   github.CheckRunState
-		ExpectedAuditJobRequest *activities.AuditJobRequest
-		ExpectedActions         []github.CheckRunAction
-	}{
-		{
+	t.Run("calls notifier with state", func(t *testing.T) {
+		ts := testsuite.WorkflowTestSuite{}
+		env := ts.NewTestWorkflowEnvironment()
+
+		env.ExecuteWorkflow(testStateReceiveWorkflow, stateReceiveRequest{
 			State: &state.Workflow{
 				Plan: &state.Job{
 					Output: jobOutput,
 					Status: state.WaitingJobStatus,
 				},
 			},
-			ExpectedCheckRunState: github.CheckRunPending,
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.InProgressJobStatus,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunPending,
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.FailedJobStatus,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunPending,
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.FailedJobStatus,
-				},
-				Result: state.WorkflowResult{
-					Status: state.CompleteWorkflowStatus,
-					Reason: state.InternalServiceError,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunFailure,
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunPending,
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-				Apply: &state.Job{
-					Output: jobOutput,
-					Status: state.WaitingJobStatus,
-					OnWaitingActions: state.JobActions{
-						Actions: []state.JobAction{
-							{
-								ID:   "Confirm",
-								Info: "confirm",
-							},
-						},
-					},
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunActionRequired,
-			ExpectedActions: []github.CheckRunAction{
-				{
-					Label:       "Confirm",
-					Description: "confirm",
-				},
-			},
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-				Apply: &state.Job{
-					Output:    jobOutput,
-					Status:    state.InProgressJobStatus,
-					StartTime: stTime,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunPending,
-			ExpectedAuditJobRequest: &activities.AuditJobRequest{
-				Root:         internalDeploymentInfo.Root,
-				Repo:         internalDeploymentInfo.Repo,
-				Revision:     internalDeploymentInfo.Commit.Revision,
-				State:        activities.AtlantisJobStateRunning,
-				StartTime:    strconv.FormatInt(stTime.Unix(), 10),
-				IsForceApply: false,
-			},
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-				Apply: &state.Job{
-					Output:    jobOutput,
-					Status:    state.FailedJobStatus,
-					StartTime: stTime,
-					EndTime:   endTime,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunPending,
-			ExpectedAuditJobRequest: &activities.AuditJobRequest{
-				Root:         internalDeploymentInfo.Root,
-				Repo:         internalDeploymentInfo.Repo,
-				Revision:     internalDeploymentInfo.Commit.Revision,
-				State:        activities.AtlantisJobStateFailure,
-				StartTime:    strconv.FormatInt(stTime.Unix(), 10),
-				EndTime:      strconv.FormatInt(endTime.Unix(), 10),
-				IsForceApply: false,
-			},
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-				Apply: &state.Job{
-					Output:    jobOutput,
-					Status:    state.FailedJobStatus,
-					StartTime: stTime,
-					EndTime:   endTime,
-				},
-				Result: state.WorkflowResult{
-					Status: state.CompleteWorkflowStatus,
-					Reason: state.InternalServiceError,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunFailure,
-			ExpectedAuditJobRequest: &activities.AuditJobRequest{
-				Root:         internalDeploymentInfo.Root,
-				Repo:         internalDeploymentInfo.Repo,
-				Revision:     internalDeploymentInfo.Commit.Revision,
-				State:        activities.AtlantisJobStateFailure,
-				StartTime:    strconv.FormatInt(stTime.Unix(), 10),
-				EndTime:      strconv.FormatInt(endTime.Unix(), 10),
-				IsForceApply: false,
-			},
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-				Apply: &state.Job{
-					Output:    jobOutput,
-					Status:    state.FailedJobStatus,
-					StartTime: stTime,
-					EndTime:   endTime,
-				},
-				Result: state.WorkflowResult{
-					Status: state.CompleteWorkflowStatus,
-					Reason: state.TimeoutError,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunTimeout,
-			ExpectedAuditJobRequest: &activities.AuditJobRequest{
-				Root:         internalDeploymentInfo.Root,
-				Repo:         internalDeploymentInfo.Repo,
-				Revision:     internalDeploymentInfo.Commit.Revision,
-				State:        activities.AtlantisJobStateFailure,
-				StartTime:    strconv.FormatInt(stTime.Unix(), 10),
-				EndTime:      strconv.FormatInt(endTime.Unix(), 10),
-				IsForceApply: false,
-			},
-		},
-		{
-			State: &state.Workflow{
-				Plan: &state.Job{
-					Output: jobOutput,
-					Status: state.SuccessJobStatus,
-				},
-				Apply: &state.Job{
-					Output:    jobOutput,
-					Status:    state.SuccessJobStatus,
-					StartTime: stTime,
-					EndTime:   endTime,
-				},
-				Result: state.WorkflowResult{
-					Status: state.CompleteWorkflowStatus,
-					Reason: state.SuccessfulCompletionReason,
-				},
-			},
-			ExpectedCheckRunState: github.CheckRunSuccess,
-			ExpectedAuditJobRequest: &activities.AuditJobRequest{
-				Root:         internalDeploymentInfo.Root,
-				Repo:         internalDeploymentInfo.Repo,
-				Revision:     internalDeploymentInfo.Commit.Revision,
-				State:        activities.AtlantisJobStateSuccess,
-				StartTime:    strconv.FormatInt(stTime.Unix(), 10),
-				EndTime:      strconv.FormatInt(endTime.Unix(), 10),
-				IsForceApply: false,
-			},
-		},
-	}
-
-	for _, c := range cases {
-		t.Run("new version", func(t *testing.T) {
-			ts := testsuite.WorkflowTestSuite{}
-			env := ts.NewTestWorkflowEnvironment()
-
-			var a = &testActivities{}
-			env.RegisterActivity(a)
-
-			env.AssertNotCalled(t, "GithubUpdateCheckRun", activities.UpdateCheckRunRequest{
-				Title: "atlantis/deploy: root",
-				State: c.ExpectedCheckRunState,
-				Repo: github.Repo{
-					Name: "hello",
-				},
-				Summary: markdown.RenderWorkflowStateTmpl(c.State),
-				ID:      1,
-				Actions: c.ExpectedActions,
-			})
-
-			if c.ExpectedAuditJobRequest != nil {
-				env.OnActivity(a.AuditJob, mock.Anything, *c.ExpectedAuditJobRequest).Return(nil)
-			}
-
-			env.ExecuteWorkflow(testStateReceiveWorkflow, stateReceiveRequest{
-				StatesToSend:   []*state.Workflow{c.State},
-				DeploymentInfo: internalDeploymentInfo,
-				ExpectedRequest: notifier.GithubCheckRunRequest{
-					Title: "atlantis/deploy: root",
-					Sha:   internalDeploymentInfo.Commit.Revision,
-					State: c.ExpectedCheckRunState,
-					Repo: github.Repo{
-						Name: "hello",
-					},
-					Summary: markdown.RenderWorkflowStateTmpl(c.State),
-					Actions: c.ExpectedActions,
-				},
-				T: t,
-			})
-
-			env.AssertExpectations(t)
-
-			err = env.GetWorkflowResult(nil)
-			assert.NoError(t, err)
+			DeploymentInfo: internalDeploymentInfo,
+			T:              t,
 		})
-	}
+
+		env.AssertExpectations(t)
+
+		var result stateReceiveResponse
+		err = env.GetWorkflowResult(&result)
+		assert.True(t, result.NotifierCalled)
+		assert.NoError(t, err)
+	})
 }
